Tratamento de exceçãoes.

+++ src/main/java/com/softstrem/dscommerce/services/ProductService.java
@@ -1,16 +1,20 @@
 package com.softstrem.dscommerce.services;
 
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.softstrem.dscommerce.dto.ProductDTO;
 import com.softstrem.dscommerce.entities.Product;
 import com.softstrem.dscommerce.repositories.ProductRepository;
+import com.softstrem.dscommerce.services.exceptions.DatabaseException;
+import com.softstrem.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -19,8 +23,8 @@ public class ProductService {
 	private ProductRepository repository;
 
 	@Transactional(readOnly = true)
-	public Optional<ProductDTO> findById(Long id) {
-		return repository.findById(id).map(product -> new ProductDTO(product));
+	public ProductDTO findById(Long id) {
+		return repository.findById(id).map(product -> new ProductDTO(product)).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
 
 	}
 
@@ -42,16 +46,29 @@ public class ProductService {
 
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		Product entity = repository.getReferenceById(id);
-		copyDtoToEntity(dto, entity);
-		entity = repository.save(entity);
-		return new ProductDTO(entity);
+		try {
+			Product entity = repository.getReferenceById(id);
+			copyDtoToEntity(dto, entity);
+			entity = repository.save(entity);
+			return new ProductDTO(entity);
+
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrato");
+		}
 
 	}
 
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+
+		try {
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial");
+		}
 
 	}
